vite-node, gc endpoint
