+++ src/lib/server/gc.ts
@@ -7,3 +7,18 @@ export const safeGC = () => {
 		console.error(e?.message);
 	}
 };
+const formatMemoryUsage = (data: number) => `${Math.round((data / 1024 / 1024) * 100) / 100} MB`;
+export const memoryUsage = () => {
+	const memoryData = process.memoryUsage();
+	const memoryUsage = {
+		rss: `${formatMemoryUsage(
+			memoryData.rss
+		)} -> Resident Set Size - total memory allocated for the process execution`,
+		heapTotal: `${formatMemoryUsage(memoryData.heapTotal)} -> total size of the allocated heap`,
+		heapUsed: `${formatMemoryUsage(
+			memoryData.heapUsed
+		)} -> actual memory used during the execution`,
+		external: `${formatMemoryUsage(memoryData.external)} -> V8 external memory`
+	};
+	return memoryUsage;
+};
