+++ src/routes/admin/gc/+server.ts
@@ -1,0 +1,14 @@
+import { memoryUsage, safeGC } from '$lib/server/gc';
+import { json } from '@sveltejs/kit';
+
+import type { RequestHandler } from './$types';
+
+export const GET: RequestHandler = async ({ url }) => {
+	const before = memoryUsage();
+	safeGC();
+	const after = memoryUsage();
+	return json({
+		before,
+		after
+	});
+};
